+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Remove 'analytics' from metatags
+
 # 18.6.1
 
 * Update body classes for govuk-frontend-v5